first_task/requirements.txt
dotenv
psycopg2-binary
pandas
sqlalchemy
unittest